backend/pyproject.toml
[project]
name = "scla-backend"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2.0",
  "alembic",
  "pydantic>=2",
  "psycopg[binary]",
  "httpx",
  "beautifulsoup4",
  "pillow",
  "python-multipart",
  "fastapi-cors"
]
[tool.setuptools]
py-modules = []
